Rename components to providers
