+++ packages/hooks/src/index.ts
@@ -1,2 +1,2 @@
-export * from './components'
 export * from './hooks'
+export * from './providers'
