+++ packages/hooks/test/wrapper.tsx
@@ -3,7 +3,7 @@ import { publicProvider, StarknetConfig } from '@starknet-react/core'
 import { Factory } from 'core'
 import { constants, RpcProvider } from 'starknet'
 
-import { Provider } from '../src/components/Provider'
+import { Provider } from '../src/providers/Provider'
 
 const provider = new RpcProvider({
   nodeUrl: 'https://starknet-mainnet.public.blastapi.io',
